# swagger.yaml
openapi: 3.0.0
info:
  title: TrailAPI
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local server
  - url: https://dist-6-505.uopnet.plymouth.ac.uk
    description: Uni Server

paths:
  /trails:
    get:
      summary: Get all trails
      operationId: get_trails
      tags:
        - Trails
      responses:
        '200':
          description: List of trails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trail'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new trail
      operationId: add_trail
      tags:
        - Trails
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrailInput'
      responses:
        '201':
          description: Trail created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trail'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /trails/{id}:
    get:
      summary: Get trail by ID
      operationId: get_trail_by_id
      tags:
        - Trails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Trail details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update trail
      operationId: update_trail
      tags:
        - Trails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrailInput'
      responses:
        '200':
          description: Trail updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trail'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete trail
      operationId: delete_trail
      tags:
        - Trails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Trail:
      type: object
      properties:
        trail_id:
          type: integer
        name:
          type: string
        difficulty:
          type: string
        description:
          type: string
        distance:
          type: number
          format: float
        elevation:
          type: number
          format: float
        hours:
          type: integer
        mins:
          type: integer
        location:
          type: object
          properties:
            city:
              type: string
            country:
              type: string
            county:
              type: string
        type:
          type: string
        owner:
          type: object
          properties:
            user_id:
              type: integer
            email:
              type: string
            role:
              type: string

    TrailInput:
      type: object
      required:
        - name
        - difficulty
        - distance
        - description
        - location_id
        - type_id
        - owner_id
      properties:
        name:
          type: string
        difficulty:
          type: string
        description:
          type: string
        distance:
          type: number
          format: float
        elevation:
          type: number
          format: float
        hours:
          type: integer
        mins:
          type: integer
        location_id:
          type: integer
        type_id:
          type: integer
        owner_id:
          type: integer
        city:
          type: string
        county:
          type: string
        country:
          type: string

    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string

    Location:
      type: object
      properties:
        location_id:
          type: integer
        city:
          type: string
        county:
          type: string
        country:
          type: string

    Type:
      type: object
      properties:
        type_id:
          type: integer
        type_name:
          type: string

    User:
      type: object
      properties:
        user_id:
          type: integer
        email:
          type: string
        role:
          type: string